+++ tests/dummy/app/controllers/ember-flexberry-dummy-suggestion-type-edit.js
@@ -39,5 +39,13 @@ export default BaseEditFormController.extend(EditFormControllerOperationsIndicat
     }
 
     return cellComponent;
+  },
+
+  actions: {
+    configurateSelectedRows(selectedRecords) {
+      if (selectedRecords.length > 0) {
+        console.log("selectedRecords.length", selectedRecords.length)
+      }
+    }
   }
 });